Popup Error Param
Added a popup param option that retains the original functionality of the script errors.

These errors have caused issues in the past for certain developers that were unable to prevent them from spamming. By using the console we can still show the error without forcing it to disrupt the gameplay, similar to gamemode.

There are advantages to the old system, especially for more advanced users, so its still an option. It would definitely be better to simply make this an actual option in the editor, so I'll probably change it to work that way eventually so the user can switch it on the fly.