Do *NOT* call this here as it will buckle us at load.
Mind this could break when running and loading a level via the editor after the fact? Possibly only edge case I can think where this matters. If that's a problem, this call should be moved to somewhere *after* otherwise this method will get called again and again and again on each merge.